send log only on error

+++ scripts/sh/sendLogToBitbucket.sh
@@ -15,19 +15,22 @@
 #   8: git repo key
 #   9: git pull request key: ${bamboo.repository.pr.key}
 
-shopt -s expand_aliases
-alias jq=C:/Users/virt7173/bamboo-home/tools/jq-win64.exe
+if [ ! -f somefile.temp ];
+then
+  shopt -s expand_aliases
+  alias jq=C:/Users/virt7173/bamboo-home/tools/jq-win64.exe
 
-# download log:
-curl -X GET --user ${1}:${2} "http://gitvyvoj.cmss.local:8086/download/${5}/build_logs/${6}.log" -o SCRM-SFCI3-JOB1-3.log
+  # download log:
+  curl -X GET --user ${1}:${2} "http://gitvyvoj.cmss.local:8086/download/${5}/build_logs/${6}.log" -o SCRM-SFCI3-JOB1-3.log
 
-attachmentUrl=$(curl -k -u ${3}:${4} --insecure \
-     -X POST 'https://gitvyvoj.cmss.local:8443/projects/${7}/repos/${8}/attachments' \
-     -H 'Content-Type: multipart/form-data;' \
-     -F 'files=@${6}.log' \
-     | jq -j '.attachments[0].url')
+  attachmentUrl=$(curl -k -u ${3}:${4} --insecure \
+       -X POST "https://gitvyvoj.cmss.local:8443/projects/${7}/repos/${8}/attachments" \
+       -H 'Content-Type: multipart/form-data;' \
+       -F "files=@${6}.log" \
+       | jq -j '.attachments[0].url')
 
-curl -k -u ${3}:${4} \
-     -H 'Content-type: application/json' \
-     -X POST 'https://gitvyvoj.cmss.local:8443/rest/api/latest/projects/${7}/repos/${8}/pull-requests/${9}/comments' \
-     -d '{"text": "Log file for last build: [${6}.log](${attachmentUrl})"}'
+  curl -k -u ${3}:${4} \
+       -H 'Content-type: application/json' \
+       -X POST "https://gitvyvoj.cmss.local:8443/rest/api/latest/projects/${7}/repos/${8}/pull-requests/${9}/comments" \
+       -d "{\"text\": \"Last build failed. See log file for more details: [${6}.log](${attachmentUrl})\"}"
+fi
